*Optimized the automatic usb port switching. It is now a lot faster on mounting if partitions are FAT/NTFS/EXT because no port switch is required on WBFS_Open() or WBFS_Close() on those partitions.
*Disabled use of custom paths on other than SD if automatic port switching is enabled. This could damage a partition and we don't want to take that risk. So if you want to enable automatic usb port switching than you have to set all paths on sd:/ first now. Also when in this mode, only sd:/ will be available for setting custom paths.
*Changed setting of usb port switching to a more failsafer way.

NOTE: To all that did already enable the automatic usb port switching and have custom paths on some usb partition. Change this please or else your usbs filesystems are in danger. You have been warned!

+++ source/usbloader/wbfs/wbfs_fat.cpp
@@ -50,7 +50,7 @@ s32 Wbfs_Fat::Open()
 
     PartitionHandle * usbHandle = DeviceHandler::Instance()->GetUSBHandle();
 
-    if(partition >= 0 && partition < usbHandle->GetPartitionTotalCount())
+    if(partition >= 0 && (int) partition < usbHandle->GetPartitionTotalCount())
     {
         if (lba == usbHandle->GetLBAStart(partition))
         {
